+++ src/components/RoadMap/index.js
@@ -3,5 +3,4 @@ export * from './Course';
 export * from './Table';
 export * from './TotalRoadMap';
 
-export { default as Main } from './Main';
 export { default as RoadMapContents } from './RoadMapContents';
